+++ common/steamcraft/common/block/BlockUraniteOre.java
@@ -13,15 +13,17 @@
  */
 package common.steamcraft.common.block;
 
-import common.steamcraft.client.core.helper.IconHelper;
-import cpw.mods.fml.relauncher.Side;
-import cpw.mods.fml.relauncher.SideOnly;
+import java.util.Random;
+
 import net.minecraft.block.Block;
 import net.minecraft.block.material.Material;
 import net.minecraft.client.renderer.texture.IconRegister;
 import net.minecraft.world.World;
 
-import java.util.Random;
+import common.steamcraft.client.core.helper.IconHelper;
+
+import cpw.mods.fml.relauncher.Side;
+import cpw.mods.fml.relauncher.SideOnly;
 
 /**
  * 
@@ -104,35 +106,35 @@ public class BlockUraniteOre extends BlockMod
 		
 		for(int l = 0; l < 6; l++)
 		{
-			double d1 = (float)i + random.nextFloat();
-			double d2 = (float)j + random.nextFloat();
-			double d3 = (float)k + random.nextFloat();
+			double d1 = i + random.nextFloat();
+			double d2 = j + random.nextFloat();
+			double d3 = k + random.nextFloat();
 		
 			if(l == 0 && !world.isBlockOpaqueCube(i, j + 1, k))
 			{
-				d2 = (double)(j + 1) + d;
+				d2 = j + 1 + d;
 			}
 			if(l == 1 && !world.isBlockOpaqueCube(i, j - 1, k))
 			{
-				d2 = (double)(j + 0) - d;
+				d2 = j + 0 - d;
 			}
 			if(l == 2 && !world.isBlockOpaqueCube(i, j, k + 1))
 			{
-				d3 = (double)(k + 1) + d;
+				d3 = k + 1 + d;
 			}
 			if(l == 3 && !world.isBlockOpaqueCube(i, j, k - 1))
 			{
-				d3 = (double)(k + 0) - d;
+				d3 = k + 0 - d;
 			}
 			if(l == 4 && !world.isBlockOpaqueCube(i + 1, j, k))
 			{
-				d1 = (double)(i + 1) + d;
+				d1 = i + 1 + d;
 			}
 			if(l == 5 && !world.isBlockOpaqueCube(i - 1, j, k))
 			{
-				d1 = (double)(i + 0) - d;
+				d1 = i + 0 - d;
 			}
-			if(d1 < (double)i || d1 > (double)(i + 1) || d2 < 0.0D || d2 > (double)(j + 1) || d3 < (double)k || d3 > (double)(k + 1))
+			if(d1 < i || d1 > i + 1 || d2 < 0.0D || d2 > j + 1 || d3 < k || d3 > k + 1)
 			{
 				world.spawnParticle("reddust", d1, d2, d3, -1.0D, 1.0D, -1.0D);
 			}
