+++ common/steamcraft/common/item/ItemModSword.java
@@ -1,8 +1,5 @@
 package common.steamcraft.common.item;
 
-import com.google.common.collect.Multimap;
-import cpw.mods.fml.relauncher.Side;
-import cpw.mods.fml.relauncher.SideOnly;
 import net.minecraft.block.Block;
 import net.minecraft.block.material.Material;
 import net.minecraft.entity.EntityLivingBase;
@@ -13,6 +10,11 @@ import net.minecraft.item.EnumAction;
 import net.minecraft.item.EnumToolMaterial;
 import net.minecraft.item.ItemStack;
 import net.minecraft.world.World;
+
+import com.google.common.collect.Multimap;
+
+import cpw.mods.fml.relauncher.Side;
+import cpw.mods.fml.relauncher.SideOnly;
 
 public class ItemModSword extends ItemModTool
 {
@@ -45,7 +47,7 @@ public class ItemModSword extends ItemModTool
 	public boolean hitEntity(ItemStack stack, EntityLivingBase living1, EntityLivingBase living2)
 	{	
 		int itemDamage = stack.getItemDamage();
-		System.out.println(weaponDamage - (int) Math.round(itemDamage * 10 / 320));
+		System.out.println(weaponDamage - Math.round(itemDamage * 10 / 320));
 		stack.damageItem(1, living2);
 		return true;
 	}
@@ -53,7 +55,7 @@ public class ItemModSword extends ItemModTool
 	@Override
 	public boolean onBlockDestroyed(ItemStack stack, World world, int i, int j, int k, int l, EntityLivingBase living)
 	{
-		if((double)Block.blocksList[l].getBlockHardness(world, i, j, k) != 0.0D)
+		if(Block.blocksList[l].getBlockHardness(world, i, j, k) != 0.0D)
 		{
 			stack.damageItem(2, living);
 		}
@@ -110,7 +112,7 @@ public class ItemModSword extends ItemModTool
 	public Multimap getItemAttributeModifiers()
 	{
 		Multimap multimap = super.getItemAttributeModifiers();
-		multimap.put(SharedMonsterAttributes.attackDamage.getAttributeUnlocalizedName(), new AttributeModifier(field_111210_e, "Weapon modifier", (double)this.weaponDamage, 0));
+		multimap.put(SharedMonsterAttributes.attackDamage.getAttributeUnlocalizedName(), new AttributeModifier(field_111210_e, "Weapon modifier", this.weaponDamage, 0));
 		return multimap;
 	}
 }
